Refactor GPT data
