+++ fast_llm/engine/training/config.py
@@ -7,7 +7,7 @@ import subprocess
 import typing
 
 from fast_llm.config import Config, Field, FieldHint, FieldUpdate, check_field, config_class, skip_valid_if_none
-from fast_llm.data.config import AbstractDataConfig
+from fast_llm.data.config import DataConfig
 from fast_llm.engine.checkpoint.config import (
     CheckpointLoadConfig,
     CheckpointSaveConfig,
@@ -330,8 +330,8 @@ class TrainerConfig(PretrainedFastLLMModelConfig, ExperimentConfig):
     schedule: ScheduleConfig = Field(
         default_factory=ScheduleConfig, desc="Configuration for the scheduling of each iteration.", hint=FieldHint.core
     )
-    data: AbstractDataConfig = Field(
-        default_factory=AbstractDataConfig,
+    data: DataConfig = Field(
+        default_factory=DataConfig,
         desc="Configuration for the dataset and model-independent preprocessing.",
         hint=FieldHint.core,
     )
